refactor: move route components from pages/ to domains/ structure
Aligns with CONVENTIONS.md 'organize by domain, not technical layer'
convention. pages/ is a Next.js file-based routing convention that doesn't
apply to React Router Data mode apps.

- src/domains/chat/chat-page.tsx
- src/domains/library/library-page.tsx
- src/domains/library/library-detail-page.tsx
- src/domains/settings/settings-tab-page.tsx
- src/components/not-found.tsx (cross-cutting, not domain-specific)
