+++ apps/web/src/components/not-found.tsx
@@ -5,7 +5,7 @@ export function NotFound() {
     <section>
       <h2>Not found</h2>
       <p>
-        The page you requested does not exist. <Link to="/conversations/new">Start a new conversation</Link>.
+        The page you requested does not exist. <Link to="/">Start a new conversation</Link>.
       </p>
     </section>
   );
