+++ apps/web/src/routes.tsx
@@ -1,10 +1,10 @@
 import { createBrowserRouter } from "react-router";
 import { App } from "./App.js";
-import { Chat } from "./pages/chat.js";
-import { Library } from "./pages/library.js";
-import { LibraryDetail } from "./pages/library-detail.js";
-import { NotFound } from "./pages/not-found.js";
-import { SettingsTab } from "./pages/settings-tab.js";
+import { ChatPage } from "./domains/chat/chat-page.js";
+import { LibraryPage } from "./domains/library/library-page.js";
+import { LibraryDetailPage } from "./domains/library/library-detail-page.js";
+import { NotFound } from "./components/not-found.js";
+import { SettingsTabPage } from "./domains/settings/settings-tab-page.js";
 
 export const router = createBrowserRouter(
   [
@@ -12,10 +12,10 @@ export const router = createBrowserRouter(
       path: "/",
       element: <App />,
       children: [
-        { index: true, element: <Chat /> },
-        { path: "settings/:tab", element: <SettingsTab /> },
-        { path: "library", element: <Library /> },
-        { path: "library/:appId", element: <LibraryDetail /> },
+        { index: true, element: <ChatPage /> },
+        { path: "settings/:tab", element: <SettingsTabPage /> },
+        { path: "library", element: <LibraryPage /> },
+        { path: "library/:appId", element: <LibraryDetailPage /> },
         { path: "*", element: <NotFound /> },
       ],
     },
